move dom-object files
