+++ src/AudioContext.js
@@ -1,8 +1,8 @@
 import utils from "./utils";
 import Configuration from "./utils/Configuration";
 import Immigration from "./utils/Immigration";
-import Event from "./Event";
-import EventTarget from "./EventTarget";
+import Event from "./dom/Event";
+import EventTarget from "./dom/EventTarget";
 import AudioBuffer from "./AudioBuffer";
 import AnalyserNode from "./AnalyserNode";
 import AudioBufferSourceNode from "./AudioBufferSourceNode";
