+++ src/dom/MediaStream.js
@@ -1,4 +1,4 @@
-import utils from "./utils";
+import utils from "../utils";
 import EventTarget from "./EventTarget";
 
 global.MediaStream = global.MediaStream || class MediaStream extends EventTarget {
