+++ test/dom/EventTarget.js
